emcp-manager/requirements.txt
flask==3.0.0
requests==2.31.0
ruamel.yaml>=0.18.0
